If statement: Transformations, validations and codegen.
Allowing globals variable and recognition of variables inside other blocks.

Starting the if statement codegen. Ran against an issue with the return logic.

Fixing the issue with the statements that were generated already. Needs more test.

Finalizando “nested if”

Modifying the readme, adding to the example and fixing a bug.